test_transferAllFrom() is hanging under OS X.  I have modified the unit test to detect the OS X platform and fail immediately rather than hanging while I get the JOURNAL_HA_BRANCH/QUADS_QUERY_BRANCH merge running.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4068 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/test/com/bigdata/io/TestFileChannelUtility.java
@@ -35,6 +35,8 @@ import java.nio.ByteBuffer;
 import java.nio.channels.FileChannel;
 import java.util.Random;
 
+import org.apache.system.SystemUtil;
+
 import com.bigdata.btree.BytesUtil;
 import com.bigdata.rawstore.Bytes;
 
@@ -348,6 +350,13 @@ public class TestFileChannelUtility extends TestCase {
      */
     public void test_transferAllFrom() throws IOException {
 
+        if(SystemUtil.isOSX()) {
+            /*
+             * FIXME For some reason, this unit test is hanging under OS X.
+             */
+            fail("Unit test hangs under OS X");
+        }
+        
         final File sourceFile = File.createTempFile("TestFileChannelUtility", getName());
 
         sourceFile.deleteOnExit();
@@ -358,7 +367,7 @@ public class TestFileChannelUtility extends TestCase {
 
         final RandomAccessFile source = new RandomAccessFile(sourceFile, "rw");
         
-        final RandomAccessFile target = new RandomAccessFile(sourceFile, "rw");
+        final RandomAccessFile target = new RandomAccessFile(targetFile, "rw");
        
         try {
             
@@ -377,6 +386,7 @@ public class TestFileChannelUtility extends TestCase {
             // write ground truth onto the file.
             FileChannelUtility.writeAll(source.getChannel(), ByteBuffer
                     .wrap(expected), 0L/* pos */);
+            target.setLength(FILE_SIZE);            
             
             // do a bunch of trials of random transfers.
             for(int trial=0; trial<1000; trial++) {
